Added missing parts for the complete implementation of the endpoin

+++ nourishify.api/IAM/Domain/Model/Commands/SignUpCommand.cs
@@ -8,6 +8,6 @@ public record SignUpCommand(
     string Phone,
     string Address,
     string PhotoUrl,
-    int RoleId,
+    long RoleId,
     string Password
     );
