+++ nourishify.api/IAM/Interfaces/REST/Resources/UserResource.cs
@@ -1,3 +1,3 @@
 namespace nourishify.api.IAM.Interfaces.REST.Resources;
 
-public record UserResource(long Id, string FirstName, string LastName, string Email, string Username, string Phone, string Address, string PhotoUrl, int RoleId);
+public record UserResource(long Id, string FirstName, string LastName, string Email, string Username, string Phone, string Address, string PhotoUrl, long RoleId);
